+++ packages/editor/editor-core/src/create-editor/create-plugins-list.ts
@@ -45,6 +45,7 @@ import {
   statusPlugin,
   gridPlugin,
   alignment,
+  editorDisabledPlugin,
 } from '../plugins';
 
 /**
@@ -62,6 +63,7 @@ export function getDefaultPluginsList(props: EditorProps = {}): EditorPlugin[] {
     widthPlugin,
     typeAheadPlugin,
     unsupportedContentPlugin,
+    editorDisabledPlugin,
   ];
 }