+++ packages/editor/editor-core/src/plugins/table/nodeviews/cell.tsx
@@ -11,8 +11,12 @@ import WithPluginState from '../../../ui/WithPluginState';
 import messages from '../ui/messages';
 import { pluginKey } from '../pm-plugins/main';
 import { toggleContextualMenu } from '../actions';
-import { TableCssClassName as ClassName } from '../types';
+import { TableCssClassName as ClassName, TablePluginState } from '../types';
 import { closestElement } from '../../../utils';
+import {
+  EditorDisabledPluginState,
+  pluginKey as editorDisabledPluginKey,
+} from '../../editor-disabled';
 
 export interface CellViewProps {
   node: PmNode;
@@ -26,6 +30,7 @@ export type CellProps = {
   forwardRef: (ref: HTMLElement | null) => void;
   withCursor: boolean;
   isContextualMenuOpen: boolean;
+  disabled: boolean;
 };
 
 class Cell extends React.Component<CellProps & InjectedIntlProps> {
@@ -42,12 +47,13 @@ class Cell extends React.Component<CellProps & InjectedIntlProps> {
       isContextualMenuOpen,
       forwardRef,
       intl: { formatMessage },
+      disabled,
     } = this.props;
     const labelCellOptions = formatMessage(messages.cellOptions);
 
     return (
       <div className={ClassName.CELL_NODEVIEW_WRAPPER} ref={forwardRef}>
-        {withCursor && (
+        {withCursor && !disabled && (
           <div className={ClassName.CONTEXTUAL_MENU_BUTTON}>
             <ToolbarButton
               selected={isContextualMenuOpen}
@@ -105,14 +111,24 @@ class CellView extends ReactNodeView {
     // so we trigger render manually to hide/show contextual menu button when `targetCellPosition` is updated
     return (
       <WithPluginState
-        plugins={{ pluginState: pluginKey }}
+        plugins={{
+          pluginState: pluginKey,
+          editorDisabledPlugin: editorDisabledPluginKey,
+        }}
         editorView={props.view}
-        render={({ pluginState }) => (
+        render={({
+          pluginState,
+          editorDisabledPlugin,
+        }: {
+          pluginState: TablePluginState;
+          editorDisabledPlugin: EditorDisabledPluginState;
+        }) => (
           <CellComponent
             forwardRef={forwardRef}
             withCursor={this.getPos() === pluginState.targetCellPosition}
-            isContextualMenuOpen={pluginState.isContextualMenuOpen}
+            isContextualMenuOpen={!!pluginState.isContextualMenuOpen}
             view={props.view}
+            disabled={(editorDisabledPlugin || {}).editorDisabled}
           />
         )}
       />