+++ packages/elements/user-picker/src/__tests__/unit/components/ClearIndicatorSpec.tsx
@@ -6,38 +6,6 @@ import { ClearIndicator } from '../../../components/ClearIndicator';
 describe('ClearIndicator', () => {
   const shallowClearIndicator = props => shallow(<ClearIndicator {...props} />);
 
-  it('should call onClearIndicatorHover onMouseEnter', () => {
-    const onClearIndicatorHover = jest.fn();
-    const component = shallowClearIndicator({
-      selectProps: { onClearIndicatorHover },
-    });
-
-    component.simulate('mouseEnter');
-
-    expect(onClearIndicatorHover).toHaveBeenCalledWith(true);
-  });
-
-  it('should call onClearIndicatorHover onMouseLeave', () => {
-    const onClearIndicatorHover = jest.fn();
-    const component = shallowClearIndicator({
-      selectProps: { onClearIndicatorHover },
-    });
-
-    component.simulate('mouseLeave');
-
-    expect(onClearIndicatorHover).toHaveBeenCalledWith(false);
-  });
-
-  it('should call onClearIndicator on unmount', () => {
-    const onClearIndicatorHover = jest.fn();
-    const component = shallowClearIndicator({
-      selectProps: { onClearIndicatorHover },
-    });
-
-    component.unmount();
-    expect(onClearIndicatorHover).toHaveBeenCalledWith(false);
-  });
-
   it('should clear value onMouseDown', () => {
     const clearValue = jest.fn();
     const component = shallowClearIndicator({ clearValue });