+++ website/src/pages/Examples/loader.tsx
@@ -27,6 +27,7 @@ type State = {
   packageId: string;
   groupId: string;
   exampleId: string;
+  examplesPath: string | undefined;
 };
 
 type ExampleLoaderProps = {
@@ -38,16 +39,18 @@ export default class ExamplesIFrame extends Component<{}, State> {
     packageId: '',
     groupId: '',
     exampleId: '',
+    examplesPath: undefined,
   };
   componentWillMount() {
     if (window) {
-      const { packageId, groupId, exampleId } = qs.parse(
+      const { packageId, groupId, exampleId, examplesPath } = qs.parse(
         window.location.search,
       );
       this.setState({
         packageId,
         groupId,
         exampleId,
+        examplesPath,
       });
     }
   }
@@ -68,17 +71,14 @@ export default class ExamplesIFrame extends Component<{}, State> {
   }
 
   render() {
-    const { examples, packageId, exampleId } = packageResolver(
+    const { example, packageId, exampleId } = packageResolver(
       this.state.groupId,
       this.state.packageId,
       this.state.exampleId,
+      this.state.examplesPath,
     );
-    if (examples && exampleId) {
-      return (
-        <ExampleLoader
-          example={fs.getById(fs.getFiles(examples.children), exampleId)}
-        />
-      );
+    if (example && exampleId) {
+      return <ExampleLoader example={example} />;
     }
 
     return (