+++ packages/editor/adf-utils/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @atlaskit/adf-utils
 
+## 5.1.6
+- [patch] [df33a8b](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/df33a8b):
+
+  - Fix block marks validation
+
+## 5.1.5
+- [patch] [d3bb11f](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/d3bb11f):
+
+  - Fixing validator for alignment marks
+
+## 5.1.4
+- [patch] [4e2a3b1](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/4e2a3b1):
+
+  - Fixes validating marks with multiple possible branch
+
 ## 5.1.3
 - [patch] [a1fb551](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/a1fb551):